Refactoring (#16)
* refactor: Create startup.cs

Move DI and configure app logic into file startup.cs

* refactor: Add client layer, add automapper, remove some helpers

Create project for CB client, add automapper instead of Model-DTO custom mapper, add logic for work with redis in DAL layer, add auth service instead of helpers of auth/jwt, fix tests

* fix: typo in project name

Fixing typo in project name, add checking finded user is not null in AuthService

* fix: generate jwt token

swap parameters in method VerifyPass

* refactor: rename client project, add dependencies in di container

Rename name of client layer project, add denepdencies, delete excess folder

* refactor: Custom Exceptions

Add custom exceptions for TradeService and LotService, rename XmlToDto method, change di lifetime to scoped

* refactor: Add BalanceCalculationService

Create BalanceCalculationService and move balance calculation logic from TradeService

* fix: Null owner balance

Fix problem in CalculateBalance that trying calculate owner balance, but its null. Fix naming in CreateTrade and CalculcateBalance methods a bit

* refactor: Remove all helpers, refactor tests

Delete all helpers, move them to services, refactor tests with those changes. Mocking "help"-services, move dublicate code into a method

---------

+++ CurrencyTrading.services/Services/LotService.cs
@@ -7,6 +7,7 @@ using System.Runtime.CompilerServices;
 using System.Transactions;
 using CurrencyTrading.services.Helpers;
 using CurrencyTrading.DAL.Helpers;
+using AutoMapper;
 
 namespace CurrencyTrading.services.Services
 {
@@ -14,24 +15,28 @@ namespace CurrencyTrading.services.Services
     {
         private readonly IUserRepository _userRepository;
         private readonly ILotRepository _lotRepository;
-        public LotService(IUserRepository userRepository, ILotRepository lotRepository)
+        private readonly IBalanceCalculationService _balanceCalculationService;
+        private readonly IMapper _mapper;
+        public LotService(IUserRepository userRepository, ILotRepository lotRepository, 
+            IBalanceCalculationService balanceCalculationService, IMapper mapper)
         {
             _userRepository = userRepository;
             _lotRepository = lotRepository;
+            _balanceCalculationService = balanceCalculationService;
+            _mapper = mapper;
         }
 
         public async Task<Lot> CreateLot(int userId, LotDTO lot)
         {
             var user = await _userRepository.GetUserAsync(userId);
-            var userLots = user.Lots?.ToList();
 
             if (lot.Type == Types.Sold)
             {
-                CheckBalances.CheckEnoughBalanceForSold(user, userLots, lot);
+                _balanceCalculationService.CheckEnoughBalanceForSold(user, _mapper.Map<Lot>(lot));
             }
             else
             {
-                CheckBalances.CheckEnoughBalanceForBuy(user, userLots, lot);
+                _balanceCalculationService.CheckEnoughBalanceForBuy(user, _mapper.Map<Lot>(lot));
             }
             Lot newLot = new Lot 
             {
@@ -51,6 +56,10 @@ namespace CurrencyTrading.services.Services
         public async Task<Lot> DeleteLot(int lotId)
         {
             var currentLot = await _lotRepository.GetLotAsync(lotId);
+            if(currentLot is null)
+            {
+                throw new LotNotFound();
+            }
             if (currentLot.Status != Statuses.Solded)
             {
                 var lot = await _lotRepository.DeleteLotAsync(lotId);
@@ -64,7 +73,12 @@ namespace CurrencyTrading.services.Services
 
         public async Task<Lot> GetLot(int lotId)
         {
-            return await _lotRepository.GetLotAsync(lotId);
+            var lot = await _lotRepository.GetLotAsync(lotId);
+            if (lot is null)
+            {
+                throw new LotNotFound();
+            }
+            return lot;
         }
 
         public async Task<ICollection<Lot>> GetLots()
@@ -74,17 +88,22 @@ namespace CurrencyTrading.services.Services
 
         public async Task<Lot> UpdateLot(int lotId,LotDTO lot,int userId)
         {
+            var updatedLot = await _lotRepository.GetLotAsync(lotId);
+
+            if (updatedLot is null)
+            {
+                throw new LotNotFound();
+            }
+
             var user = await _userRepository.GetUserAsync(userId);
-            var userLots = user.Lots.ToList();
             if (lot.Type == Types.Sold)
             {
-                CheckBalances.CheckEnoughBalanceForSold(user, userLots, lot);
+                _balanceCalculationService.CheckEnoughBalanceForSold(user, _mapper.Map<Lot>(lot));
             }
             else
             {
-                CheckBalances.CheckEnoughBalanceForBuy(user, userLots, lot);
+                _balanceCalculationService.CheckEnoughBalanceForBuy(user, _mapper.Map<Lot>(lot));
             }
-            var updatedLot = await _lotRepository.GetLotAsync(lotId);
 
             var newLot = UpdateEntityHelper.updateEntity(lot,updatedLot);
             if (updatedLot.Status == Statuses.Solded)
